add user authorization

+++ base/views.py
@@ -1,10 +1,13 @@
-from turtle import end_fill
-from django.db.models import Q
 from django.shortcuts import render, redirect
+from django.db.models import Q
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth.models import User
+from django.contrib.auth import authenticate, login, logout
+from django.contrib import messages
 from django.http import HttpResponse
-from urllib3 import encode_multipart_formdata
 from .models import Room, Topic
 from .forms import RoomForm
+from django.contrib.auth.forms import UserCreationForm
 
 # rooms = [
 #     {"id": 1, "name": "Learn python"},
@@ -42,6 +45,7 @@ def room(req, pk):
     context = {"room": room}
     return render(req, 'base/room.html', context)
 
+@login_required(login_url='login')
 def createRoom(req):
     form = RoomForm()
     if req.method == 'POST':
@@ -53,9 +57,13 @@ def createRoom(req):
     context = {'form': form}
     return render(req, 'base/room_form.html', context)
 
+@login_required(login_url='login')
 def updateRoom(req, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance = room)
+
+    if req.user != room.host: #also add to check for Super user
+        return HttpResponse("You're not allowed here")
     if req.method == 'POST':
         form = RoomForm(req.POST, instance=room)
         if form.is_valid:
@@ -63,9 +71,14 @@ def updateRoom(req, pk):
             return redirect('home')
     context = {'form': form}
     return render(req, 'base/room_form.html', context)
-    
+
+@login_required(login_url='login')
 def deleteRoom(req, pk):
     room = Room.objects.get(id=pk)
+
+    if req.user != room.host: #also add to check for Super user
+        return HttpResponse("You're not allowed here")
+
     if req.method == 'POST':
         room.delete()
         return redirect('home')
@@ -73,5 +86,51 @@ def deleteRoom(req, pk):
     return render(req, 'base/delete.html', context)
 
 def loginView(req):
-    context={}
+    page = 'login'
+    if req.user.is_authenticated:
+        return redirect('home')
+
+    if req.method == 'POST':
+        username = req.POST.get('username').lower()
+        password=req.POST.get('password')
+        
+        # check user existance
+        try:
+            user=User.objects.get(username=username)
+        except:
+            messages.error(req, 'Username or password is incorrect')
+
+        user = authenticate(req, username=username, password=password)
+
+        # login the user
+        if user is not None:
+            login(req, user) #this creates session
+            return redirect('home')
+        else:
+            messages.error(req, 'Username or password is incorrect')
+
+    context={'page': page}
+    return render(req, 'base/login_register.html', context)
+
+def logoutUser(req):
+    logout(req)
+    return redirect('login')
+
+def registerView(req):
+    form = UserCreationForm()
+    
+    if req.method=="POST":
+        form = UserCreationForm(req.POST)
+        try:
+            if form.is_valid:
+                user = form.save(commit=False)
+                user.username = user.username.lower()
+                user.save()
+                login(req, user) 
+                return redirect('home')
+            else:
+                messages.error(req, 'An error occured during registration')
+        except:
+            messages.error(req, "An error occured during registration")
+    context={"form": form}
     return render(req, 'base/login_register.html', context)
